requirements.txt
psycopg2-binary
asyncpg
python-dotenv